requirements.txt
asgiref==3.5.0
attrs==20.3.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
check50==3.3.5
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jellyfish==0.9.0
Jinja2==3.0.3
lib50==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
pexpect==4.8.0
psycopg2==2.9.3
ptyprocess==0.7.0
pycparser==2.21
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
submit50==3.1.0
termcolor==1.1.0
urllib3==1.26.8
whitenoise==6.0.0
